Always default to wall clock time-telling
Now that we have a `start` parameter and timeouts are anchored
in time overtly in their repr, using a monotonic clock by
default is more likely to repr to something which is either
unintuitive or does not evaluate back to the same timeout.

This way provides a more consistent behavior across more
Python implementations, for what it might be worth.

Many modern computers are getting their time synchronized
over the network or in some other reliable enough manner,
often enough for big jumps to be unlikely.

This code takes care of failing sensibly (zero time left
in the timeout) if the clock ever does jump backwards.